rename: UserName.tsx 이름 변경 #31
- 기존에 닉네임 변경을 하기 위한 파일인 UserNameEdit.tsx 를 UserName.tsx 파일로 변경했습니다
- 그에 따라 index.ts도 변경되었습니다
- 추가로 ProfileEdit.tsx 파일의 필요없는 주석도 삭제했습니다
