+++ src/components/user/index.ts
@@ -1,5 +1,5 @@
-import UserNameEdit from "./UserNameEdit";
+import UserName from "./UserName";
 import UserCategoryBtn from "./UserCategoryBtn";
 import UserList from "./UserList";
 
-export { UserNameEdit, UserCategoryBtn, UserList };
+export { UserName, UserCategoryBtn, UserList };
